+++ tests/testthat/test-US_stateConversion.R
@@ -1,48 +1,12 @@
-# TESTING: http://r-pkgs.had.co.nz/tests.html
-#   Vital to package development
-#   Part of workflow, makes sure it works
-#   Automating tests (aka unit testing)
-#   Fewer bugs, better code structure, easier restarts, robust code
-#   R is different from other languages -- functional than OO
-
-# Create a file that tests when check runs R CMD check 
-# vector of 10-20 known names/ codes/ FIPS gets properly converted into known alternatives
-# works for territories, not just states
-# handles non-matching names/ codes/ FIPS by returning na
-
-library(testthat)
-library(MazamaSpatialUtils)
-
-setSpatialDataDir('~/Data/Spatial')
-
-# ?US_stateCodes
-# ?US_stateConversion
-
-# US_stateCodeToName US_stateCodeToFIPS
-# US_stateFIPSToName US_stateFIPSToCode
-# US_stateNameToCode US_stateNameToFIPS
-# US_stateFIPSToCode
-
-# Try to set up spatial data. Skip if fails.  
-spatialDataDir <- try(getSpatialDataDir(), silent = TRUE)
-
-
-testthat::context("US_stateCodes()")
-
-#test_that("Warns when there are multiple states for a code",{
-#  expect_warning(US_stateCodeToName("HI"), "Hawaii or Howland Island?")
-#  expect_warning(US_stateCodeToName("MI"), "Michigan or Midway Islands?")
-#})
-
-
 # ----- US_stateCodeToName ----------------------------------------------------------------
 
 testthat::context("US_stateCodeToName()")
 
 test_that("Returns expected output", {
   expect_equal(US_stateCodeToName("VT"), "Vermont")
-  expect_equal(US_stateCodeToName(c("SD", "NY", "WA", "CA")),c("South Dakota", "New York", "Washington", "California"))
-  })
+  expect_equal(US_stateCodeToName(c("SD", "NY", "WA", "CA")),
+               c("South Dakota", "New York", "Washington", "California"))
+})
 
 # ----- US_stateCodeToFIPS ----------------------------------------------------------------
 
@@ -50,7 +14,8 @@ testthat::context("US_stateCodeToFIPS()")
 
 test_that("Returns expected output", {
   expect_equal(US_stateCodeToFIPS("AZ"), "04")
-  expect_equal(US_stateCodeToFIPS(c("CO", "CT", "FL", "ID")),c("08", "09", "12", "16"))
+  expect_equal(US_stateCodeToFIPS(c("CO", "CT", "FL", "ID")),
+               c("08", "09", "12", "16"))
 })
 
 # ----- US_stateFIPSToName ----------------------------------------------------------------
@@ -59,7 +24,8 @@ testthat::context("US_stateFIPSToName()")
 
 test_that("Returns expected output", {
   expect_equal(US_stateFIPSToName("17"), "Illinois")
-  expect_equal(US_stateFIPSToName(c("79", "48", "41", "01")),c("Wake Island", "Texas", "Oregon", "Alabama"))
+  expect_equal(US_stateFIPSToName(c("55", "48", "41", "01")),
+               c("Wisconsin","Texas", "Oregon", "Alabama"))
 })
 
 # ----- US_stateFIPSToCode ----------------------------------------------------------------
@@ -68,7 +34,8 @@ testthat::context("US_stateFIPSToCode()")
 
 test_that("Returns expected output", {
   expect_equal(US_stateFIPSToCode("17"), "IL")
-  expect_equal(US_stateFIPSToCode(c("79", "48", "41", "01")),c("WI", "TX", "OR", "AL"))
+  expect_equal(US_stateFIPSToCode(c("55", "48", "41", "01")),
+               c("WI", "TX", "OR", "AL"))
 })
 
 # ----- US_stateNameToCode ----------------------------------------------------------------
@@ -77,7 +44,8 @@ testthat::context("US_stateNameToCode()")
 
 test_that("Returns expected output", {
   expect_equal(US_stateNameToCode("Rhode Island"), "RI")
-  expect_equal(US_stateNameToCode(c("Indiana", "Kansas", "Maine", "New Hampshire")),c("IN", "KS", "ME", "NH"))
+  expect_equal(US_stateNameToCode(c("Indiana", "Kansas", "Maine", "New Hampshire")),
+               c("IN", "KS", "ME", "NH"))
 })
 
 # ----- US_stateNameToFIPS ----------------------------------------------------------------
@@ -86,5 +54,6 @@ testthat::context("US_stateNameToFIPS()")
 
 test_that("Returns expected output", {
   expect_equal(US_stateNameToFIPS("Utah"), "49")
-  expect_equal(US_stateNameToFIPS(c("South Carolina", "Puerto Rico", "New Mexico", "Georgia")),c("45", "72", "35", "13"))
+  expect_equal(US_stateNameToFIPS(c("South Carolina", "Puerto Rico", "New Mexico", "Georgia")),
+               c("45", "72", "35", "13"))
 })